refactor: extract 9 inner classes from TransformAnimator.java (616→373 lines)
Extract three groups of inner classes into separate package-private files:
- OrientationData.java: 6-class orientation data hierarchy
- SmoothPositionAnimations.java: Hermite-interpolation animation classes
- PlaceAnimation.java: spatial placement animation

Fix data.subject → data.getSubject() for cross-class access.
Fix TransformAnimatorExtractionTest source dir resolution for Maven Surefire.
Remove unused imports (HermiteCubic, ForwardAndUpGuide).
All 401 tests pass (342 existing + 59 new).
